Refactor utility methods into classes
